pyproject.toml
[tool.poetry]
name = "pension-schemes"
version = "0.1.0"
description = ""
authors = ["Sherwin Varghese <sherwin.varghese@imperial.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.34.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
numpy = "^1.26.4"
sentence-transformers = "^3.0.1"
groq = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
